Add ability to send a request for a failed task
For each failed task, you have the possibility to configure a way to
send a request to the caller to notify a task, and therefore the
workflow, failed

For now the data in the request will be the last serverresponse of the
failed task.

This is done by specifying a `Request` object in the `generateRexTask`
method.