[NCL-9060] Workaround issue where dependency tasks gone
Rex has an issue where the dependency tasks are deleted too eagerly if
the env-driver-complete task has no dependants.

This commit adds an edge between env-driver-complete and repo-seal so
that env-driver-complete task has a dependant. It is then not deleted,
which means that the dependency tasks are also not deleted.

We want the dependency tasks to still be available to grab their results
after the final task is finished. The results are used to send the final
result to the caller.